breakup Sql file into modules
