+++ core/jvm/src/main/scala/zio/sql/newtypes.scala
@@ -1,0 +1,15 @@
+package zio.sql
+
+trait NewtypesModule {
+
+  type ColumnName = String
+  type TableName  = String
+
+
+  sealed case class FunctionName(name: String) extends Renderable {
+    override private[zio] def renderBuilder(builder: StringBuilder, mode: RenderMode): Unit = {
+      val _ = builder.append(name)
+    }
+  }
+
+}
